포트앱, 포스트앱 합침
