--- django_app/photo/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class PhotoConfig(AppConfig):
-    name = 'photo'
