--- django_app/photo/views.py
@@ -1,8 +1,0 @@
-from .models import Photo
-from .serializers import Photoserializer
-from rest_framework import generics
-
-
-class PhotoList(generics.ListCreateAPIView):
-    queryset = Photo.objects.all()
-    serializer_class = Photoserializer
