+++ django_app/photodiary/settings.py
@@ -57,7 +57,6 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'storages',
     'member',
-    'photo',
     'post',
     'rest_framework',
     'rest_framework.authtoken',
@@ -172,8 +171,8 @@ AWS_S3_CUSTOM_DOMAIN = '%s.s3.amazonaws.com' % AWS_STORAGE_BUCKET_NAME
 
 STATICFILES_LOCATION = 'static'
 STATIC_URL = "https://%s/%s/" % (AWS_S3_CUSTOM_DOMAIN, STATICFILES_LOCATION)
-STATICFILES_STORAGE = 'photo.custom_storages.StaticStorage'
+STATICFILES_STORAGE = 'post.custom_storages.StaticStorage'
 
 MEDIAFILES_LOCATION = 'media'
 MEDIA_URL = "https://%s/%s/" % (AWS_S3_CUSTOM_DOMAIN, MEDIAFILES_LOCATION)
-DEFAULT_FILE_STORAGE = 'photo.custom_storages.MediaStorage'
+DEFAULT_FILE_STORAGE = 'post.custom_storages.MediaStorage'
