+++ django_app/post/serializers.py
@@ -1,5 +1,6 @@
 from rest_framework.serializers import ModelSerializer
-from .models import Post
+from .models import Post, Photo
+
 
 
 class Postserializer(ModelSerializer):
@@ -7,3 +8,11 @@ class Postserializer(ModelSerializer):
         model = Post
         fields = ('title', 'author', 'content',
                   'modified_date', 'created_date')
+
+
+
+class Photoserializer(ModelSerializer):
+    class Meta:
+        model = Photo
+        fields = ('author',
+                  'modified_date', 'created_date', 'image', 'post')
