+++ django_app/post/views.py
@@ -1,7 +1,6 @@
-from .models import Post
-from .serializers import Postserializer
 from rest_framework import generics
-from member.models import MyUser
+from .models import Photo
+from .serializers import Photoserializer, Postserializer
 
 
 class PostList(generics.ListCreateAPIView):
@@ -10,4 +9,16 @@ class PostList(generics.ListCreateAPIView):
     def get_queryset(self):
         print(self.request.user.pk)
         return self.request.user.post_set.all()
+
+
+
+
+class PhotoList(generics.ListCreateAPIView):
+    serializer_class = Photoserializer
+
+
+    def get_queryset(self):
+        print(self.request.user.pk)
+        return self.request.user.post_set.all()
+
